+++ src/test/java/roomescape/MissionStepTest.java
@@ -9,16 +9,12 @@ import java.lang.reflect.Field;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.time.LocalDate;
-import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
@@ -168,219 +164,6 @@ class MissionStepTest {
 
         Integer countAfterDelete = jdbcTemplate.queryForObject("SELECT count(1) from reservation", Integer.class);
         assertThat(countAfterDelete).isEqualTo(0);
-    }
-
-    @Test
-    @DisplayName("예약 생성 시, name이 null이면 예외가 발생한다.")
-    void validateReservationWithNullName() {
-        Map<String, Object> params = new HashMap<>();
-        params.put("name", null);
-        params.put("date", TOMORROW_DATE);
-        params.put("timeId", 1);
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/reservations")
-                .then().log().all()
-                .statusCode(400);
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {"", " "})
-    @DisplayName("예약 생성 시, name이 유효하지 않으면 예외가 발생한다.")
-    void validateReservationWithNameFormat(final String name) {
-        Map<String, Object> params = new HashMap<>();
-        params.put("name", name);
-        params.put("date", TOMORROW_DATE);
-        params.put("timeId", 1);
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/reservations")
-                .then().log().all()
-                .statusCode(400);
-    }
-
-    @Test
-    @DisplayName("예약 생성 시, date가 null이면 예외가 발생한다.")
-    void validateReservationWithNullDate() {
-        Map<String, Object> params = new HashMap<>();
-        params.put("name", "브라운");
-        params.put("date", null);
-        params.put("timeId", 1);
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/reservations")
-                .then().log().all()
-                .statusCode(400);
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {"", " ", "abc"})
-    @DisplayName("예약 생성 시, date의 형식이 올바르지 않으면 예외가 발생한다.")
-    void validateReservationWithDateFormat(final String date) {
-        Map<String, Object> params = new HashMap<>();
-        params.put("name", "브라운");
-        params.put("date", date);
-        params.put("timeId", 1);
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/reservations")
-                .then().log().all()
-                .statusCode(400);
-    }
-
-    @Test
-    @DisplayName("예약 생성 시, date가 이미 지난 날짜면 예외가 발생한다.")
-    void validateReservationWithPastDate() {
-        Map<String, Object> params = new HashMap<>();
-        params.put("name", "브라운");
-        params.put("date", "2020-10-10");
-        params.put("timeId", 1);
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/reservations")
-                .then().log().all()
-                .statusCode(400);
-    }
-
-    @Test
-    @DisplayName("예약 생성 시, date는 오늘이고 time은 이미 지난 시간이면 예외가 발생한다.")
-    void validateReservationWithTodayPastTime() {
-        Map<String, String> params = new HashMap<>();
-        params.put("startAt", LocalTime.now().minusHours(1).format(DateTimeFormatter.ofPattern("HH:mm")));
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/times")
-                .then().log().all()
-                .statusCode(201);
-
-        Map<String, Object> reservationParams = new HashMap<>();
-        reservationParams.put("name", "브라운");
-        reservationParams.put("date", LocalDate.now().format(DateTimeFormatter.ISO_DATE));
-        reservationParams.put("timeId", 2);
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(reservationParams)
-                .when().post("/reservations")
-                .then().log().all()
-                .statusCode(400);
-    }
-
-    @Test
-    @DisplayName("예약 생성 시, timeId가 null이면 예외가 발생한다.")
-    void validateReservationWithNullTimeId() {
-        Map<String, Object> params = new HashMap<>();
-        params.put("name", "브라운");
-        params.put("date", TOMORROW_DATE);
-        params.put("timeId", null);
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/reservations")
-                .then().log().all()
-                .statusCode(400);
-    }
-
-    @Test
-    @DisplayName("예약 생성 시, timeId 값으로 찾을 수 있는 시간이 없으면 예외가 발생한다.")
-    void validateReservationWithTimeIdNotFound() {
-        Map<String, Object> params = new HashMap<>();
-        params.put("name", "브라운");
-        params.put("date", TOMORROW_DATE);
-        params.put("timeId", 2);
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/reservations")
-                .then().log().all()
-                .statusCode(400);
-    }
-
-    @Test
-    @DisplayName("예약 생성 시, 해당 날짜와 시간에 예약 내역이 있으면 예외가 발생한다.")
-    void validateReservationWithDuplicatedDateAndTime() {
-        Map<String, Object> params = new HashMap<>();
-        params.put("name", "브라운");
-        params.put("date", TOMORROW_DATE);
-        params.put("timeId", 1);
-
-        RestAssured.given().log().all()
-            .contentType(ContentType.JSON)
-            .body(params)
-            .when().post("/reservations")
-            .then().log().all()
-            .statusCode(201);
-
-        RestAssured.given().log().all()
-            .contentType(ContentType.JSON)
-            .body(params)
-            .when().post("/reservations")
-            .then().log().all()
-            .statusCode(400);
-    }
-
-    @Test
-    @DisplayName("시간 생성 시, startAt 값이 null이면 예외가 발생한다.")
-    void validateTimeCreateWithNull() {
-        Map<String, Object> params = new HashMap<>();
-        params.put("startAt", null);
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/times")
-                .then().log().all()
-                .statusCode(400);
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {"", " ", "10:89"})
-    @DisplayName("시간 생성 시, startAt 값의 형식이 올바르지 않으면 예외가 발생한다.")
-    void validateTimeCreateWithEmpty(final String startAt) {
-        Map<String, Object> params = new HashMap<>();
-        params.put("startAt", startAt);
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/times")
-                .then().log().all()
-                .statusCode(400);
-    }
-
-    @Test
-    @DisplayName("시간 삭제 시, 해당 시간을 참조하는 예약이 있으면 예외가 발생한다.")
-    void validateTimeDelete() {
-        Map<String, Object> params = new HashMap<>();
-        params.put("name", "브라운");
-        params.put("date", TOMORROW_DATE);
-        params.put("timeId", 1);
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/reservations")
-                .then().log().all()
-                .statusCode(201);
-
-        RestAssured.given().log().all()
-                .when().delete("/times/1")
-                .then().log().all()
-                .statusCode(400);
     }
 
     @Test
